requirements.txt
attrs
gym[atari,accept-rom-license]
gymnasium[atari,accept-rom-license]
h5py
keras
matplotlib
numpy
pillow
protobuf>=3.0
pydot-ng
scipy
semver
tensorflow>=2.0.0
-e .
